+++ HISTORY.rst
@@ -2,7 +2,13 @@
 History
 =======
 
-0.1.0.1 (2017-08-24)
+0.1.0.2 (2017-09-03)
+------------------
+
+* qSigmoid now accept a numpy array as input.
+* Made some tests for qSigmoid.
+
+0.1.0.1 (2017-08-31)
 ------------------
 
 * First release on PyPI.